# Enable 1x1 convolution optimization